requirements.txt
python-telegram-bot
g4f[all]